+++ cmd/zetatool/cctx/inbound.go
@@ -128,6 +128,7 @@ func (c *TrackingDetails) btcInboundBallotIdentifier(ctx *context.Context) error
 	if err != nil {
 		return fmt.Errorf("failed to get chain params: %w", err)
 	}
+	createConfirmationParamsIfAbsent(chainParams)
 
 	cctxIdentifier, isConfirmed, err := zetatoolchains.BitcoinBallotIdentifier(
 		ctx,
@@ -137,7 +138,7 @@ func (c *TrackingDetails) btcInboundBallotIdentifier(ctx *context.Context) error
 		inboundHash,
 		inboundChain.ChainId,
 		zetaChainID,
-		chainParams.ConfirmationCount,
+		chainParams.InboundConfirmationSafe(),
 	)
 	if err != nil {
 		return fmt.Errorf("failed to get bitcoin ballot identifier: %w", err)
@@ -161,6 +162,7 @@ func (c *TrackingDetails) evmInboundBallotIdentifier(ctx *context.Context) error
 	if err != nil {
 		return fmt.Errorf("failed to get chain params: %w", err)
 	}
+	createConfirmationParamsIfAbsent(chainParams)
 
 	evmClient, err := zetatoolchains.GetEvmClient(ctx, inboundChain)
 	if err != nil {
@@ -173,7 +175,7 @@ func (c *TrackingDetails) evmInboundBallotIdentifier(ctx *context.Context) error
 	}
 	// Signer is unused
 	zetaEvmClient := zetaevmclient.New(evmClient, ethtypes.NewLondonSigner(tx.ChainId()))
-	isConfirmed, err := zetaEvmClient.IsTxConfirmed(goCtx, inboundHash, chainParams.ConfirmationCount)
+	isConfirmed, err := zetaEvmClient.IsTxConfirmed(goCtx, inboundHash, chainParams.InboundConfirmationSafe())
 	if err != nil {
 		return fmt.Errorf("unable to confirm tx: %w", err)
 	}
@@ -310,6 +312,7 @@ func (c *TrackingDetails) solanaInboundBallotIdentifier(ctx *context.Context) er
 	if err != nil {
 		return fmt.Errorf("failed to get chain params: %w", err)
 	}
+	createConfirmationParamsIfAbsent(chainParams)
 
 	gatewayID, _, err := solanacontracts.ParseGatewayWithPDA(chainParams.GatewayAddress)
 	if err != nil {
@@ -370,4 +373,16 @@ func compareAddress(a string, b string) bool {
 	lowerA := strings.ToLower(a)
 	lowerB := strings.ToLower(b)
 	return strings.EqualFold(lowerA, lowerB)
+}
+
+// createConfirmationParamsIfAbsent sets the confirmation params if they are not already set
+// TODO: Remove this once the confirmation migration is done
+// https://github.com/zeta-chain/node/issues/3466
+func createConfirmationParamsIfAbsent(chainParams *types.ChainParams) {
+	if chainParams != nil && chainParams.ConfirmationParams == nil {
+		chainParams.ConfirmationParams = &types.ConfirmationParams{
+			SafeInboundCount:  chainParams.ConfirmationCount,
+			SafeOutboundCount: chainParams.ConfirmationCount,
+		}
+	}
 }