+++ x/crosschain/types/cmd_cctxs.go
@@ -291,6 +291,10 @@ func newCmdCCTX(
 			Sender:       sender,
 			CoinType:     coin.CoinType_Cmd,
 			ObservedHash: inboundHash,
+			// irrelevant to observer voting, set it to success by default
+			Status: InboundStatus_SUCCESS,
+			// any inbound initiated from ZetaChain is deemed safely confirmed
+			ConfirmationMode: ConfirmationMode_SAFE,
 		},
 		OutboundParams: []*OutboundParams{
 			{
@@ -304,6 +308,9 @@ func newCmdCCTX(
 				GasPrice:       medianGasPrice,
 				GasPriorityFee: priorityFee,
 				TssPubkey:      tssPubKey,
+				// use SAFE confirmation mode as default value.
+				// zetaclient should ALWAYS use SAFE confirmation mode to confirm a CMD tx.
+				ConfirmationMode: ConfirmationMode_SAFE,
 			},
 		},
 	}