+++ zetaclient/chains/base/observer.go
@@ -212,15 +212,6 @@ func (ob *Observer) WithLastBlock(lastBlock uint64) *Observer {
 	return ob
 }
 
-// IsBlockConfirmed checks if the given block number is confirmed.
-//
-// Note: block 100 is confirmed if the last block is 100 and confirmation count is 1.
-func (ob *Observer) IsBlockConfirmed(blockNumber uint64) bool {
-	lastBlock := ob.LastBlock()
-	confBlock := blockNumber + ob.chainParams.ConfirmationCount - 1
-	return lastBlock >= confBlock
-}
-
 // LastBlockScanned get last block scanned (not necessarily caught up with the chain; could be slow/paused).
 func (ob *Observer) LastBlockScanned() uint64 {
 	height := atomic.LoadUint64(&ob.lastBlockScanned)