+++ zetaclient/chains/bitcoin/observer/inbound.go
@@ -12,63 +12,65 @@ import (
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 
-	"github.com/zeta-chain/node/pkg/coin"
 	"github.com/zeta-chain/node/x/crosschain/types"
 	"github.com/zeta-chain/node/zetaclient/chains/bitcoin/common"
+	"github.com/zeta-chain/node/zetaclient/config"
 	"github.com/zeta-chain/node/zetaclient/logs"
 	"github.com/zeta-chain/node/zetaclient/zetacore"
 )
 
 // ObserveInbound observes the Bitcoin chain for inbounds and post votes to zetacore
-// TODO(revamp): simplify this function into smaller functions
 func (ob *Observer) ObserveInbound(ctx context.Context) error {
-	// get and update latest block height
-	currentBlock, err := ob.rpc.GetBlockCount(ctx)
-	if err != nil {
-		return fmt.Errorf("observeInboundBTC: error getting block number: %s", err)
-	}
-	if currentBlock < 0 {
-		return fmt.Errorf("observeInboundBTC: block number is negative: %d", currentBlock)
-	}
+	logger := ob.Logger().Inbound.With().Str(logs.FieldMethod, "observe_inbound").Logger()
 
-	// 0 will be returned if the node is not synced
-	if currentBlock == 0 {
-		ob.nodeEnabled.Store(false)
-		ob.logger.Inbound.Debug().Err(err).Msg("WatchInbound: Bitcoin node is not enabled")
-		return nil
-	}
-
-	ob.nodeEnabled.Store(true)
-
-	// #nosec G115 checked positive
-	lastBlock := uint64(currentBlock)
-	if lastBlock < ob.LastBlock() {
-		return fmt.Errorf(
-			"observeInboundBTC: block number should not decrease: current %d last %d",
-			currentBlock,
-			ob.LastBlock(),
-		)
-	}
-	ob.WithLastBlock(lastBlock)
-
-	// check confirmation
-	blockNumber := ob.LastBlockScanned() + 1
-	if !ob.IsBlockConfirmed(blockNumber) {
-		return nil
-	}
-
-	// query incoming gas asset to TSS address
-	// #nosec G115 always in range
-	res, err := ob.GetBlockByNumberCached(ctx, int64(blockNumber))
-	if err != nil {
-		ob.logger.Inbound.Error().Err(err).Msgf("observeInboundBTC: error getting bitcoin block %d", blockNumber)
+	// keep last block up-to-date
+	if err := ob.updateLastBlock(ctx); err != nil {
 		return err
 	}
-	ob.logger.Inbound.Info().Msgf("observeInboundBTC: block %d has %d txs, current block %d, last scanned %d",
-		blockNumber, len(res.Block.Tx), currentBlock, blockNumber-1)
 
-	// add block header to zetacore
-	if len(res.Block.Tx) > 1 {
+	// get the block range to scan
+	startBlock, endBlock := ob.GetScanRangeInboundSafe(config.MaxBlocksPerScan)
+	if startBlock >= endBlock {
+		return nil
+	}
+
+	// observe inbounds for the block range [startBlock, endBlock-1]
+	lastScannedNew, err := ob.observeInboundInBlockRange(ctx, startBlock, endBlock-1)
+	if err != nil {
+		logger.Error().
+			Err(err).
+			Uint64("from", startBlock).
+			Uint64("to", endBlock-1).
+			Msg("error observing inbounds in block range")
+	}
+
+	// save last scanned block to both memory and db
+	if lastScannedNew > ob.LastBlockScanned() {
+		logger.Info().Uint64("from", startBlock).Uint64("to", lastScannedNew).Msg("observed blocks for inbounds")
+		if err := ob.SaveLastBlockScanned(lastScannedNew); err != nil {
+			return errors.Wrapf(err, "unable to save last scanned Bitcoin block %d", lastScannedNew)
+		}
+	}
+
+	return nil
+}
+
+// observeInboundInBlockRange observes inbounds for given block range [startBlock, toBlock (inclusive)]
+// It returns the last successfully scanned block height, so the caller knows where to resume next time
+func (ob *Observer) observeInboundInBlockRange(ctx context.Context, startBlock, toBlock uint64) (uint64, error) {
+	for blockNumber := startBlock; blockNumber <= toBlock; blockNumber++ {
+		// query incoming gas asset to TSS address
+		// #nosec G115 always in range
+		res, err := ob.GetBlockByNumberCached(ctx, int64(blockNumber))
+		if err != nil {
+			// we have to re-scan this block next time
+			return blockNumber - 1, errors.Wrapf(err, "error getting bitcoin block %d", blockNumber)
+		}
+
+		if len(res.Block.Tx) <= 1 {
+			continue
+		}
+
 		// filter incoming txs to TSS address
 		tssAddress := ob.TSSAddressString()
 
@@ -83,10 +85,8 @@ func (ob *Observer) ObserveInbound(ctx context.Context) error {
 			ob.netParams,
 		)
 		if err != nil {
-			ob.logger.Inbound.Error().
-				Err(err).
-				Msgf("observeInboundBTC: error filtering incoming txs for block %d", blockNumber)
-			return err // we have to re-scan this block next time
+			// we have to re-scan this block next time
+			return blockNumber - 1, errors.Wrapf(err, "error filtering incoming txs for block %d", blockNumber)
 		}
 
 		// post inbound vote message to zetacore
@@ -95,115 +95,15 @@ func (ob *Observer) ObserveInbound(ctx context.Context) error {
 			if msg != nil {
 				_, err = ob.PostVoteInbound(ctx, msg, zetacore.PostVoteInboundExecutionGasLimit)
 				if err != nil {
-					return errors.Wrapf(err, "error PostVoteInbound") // we have to re-scan this block next time
+					// we have to re-scan this block next time
+					return blockNumber - 1, errors.Wrapf(err, "error posting inbound vote for tx %s", event.TxHash)
 				}
 			}
 		}
 	}
 
-	// save last scanned block to both memory and db
-	if err := ob.SaveLastBlockScanned(blockNumber); err != nil {
-		ob.logger.Inbound.Error().
-			Err(err).
-			Msgf("observeInboundBTC: error writing last scanned block %d to db", blockNumber)
-	}
-
-	return nil
-}
-
-// ProcessInboundTrackers processes inbound trackers
-// TODO(revamp): move inbound tracker logic in a specific file
-func (ob *Observer) ProcessInboundTrackers(ctx context.Context) error {
-	trackers, err := ob.ZetacoreClient().GetInboundTrackersForChain(ctx, ob.Chain().ChainId)
-	if err != nil {
-		return err
-	}
-
-	for _, tracker := range trackers {
-		ob.logger.Inbound.Info().
-			Str("tracker.hash", tracker.TxHash).
-			Str("tracker.coin-type", tracker.CoinType.String()).
-			Msgf("checking tracker")
-		ballotIdentifier, err := ob.CheckReceiptForBtcTxHash(ctx, tracker.TxHash, true)
-		if err != nil {
-			return err
-		}
-		ob.logger.Inbound.Info().
-			Str("inbound.chain", ob.Chain().Name).
-			Str("inbound.ballot", ballotIdentifier).
-			Str("inbound.coin-type", coin.CoinType_Gas.String()).
-			Msgf("Vote submitted for inbound Tracker")
-	}
-
-	return nil
-}
-
-// CheckReceiptForBtcTxHash checks the receipt for a btc tx hash
-func (ob *Observer) CheckReceiptForBtcTxHash(ctx context.Context, txHash string, vote bool) (string, error) {
-	hash, err := chainhash.NewHashFromStr(txHash)
-	if err != nil {
-		return "", err
-	}
-
-	tx, err := ob.rpc.GetRawTransactionVerbose(ctx, hash)
-	if err != nil {
-		return "", err
-	}
-
-	blockHash, err := chainhash.NewHashFromStr(tx.BlockHash)
-	if err != nil {
-		return "", err
-	}
-
-	blockVb, err := ob.rpc.GetBlockVerbose(ctx, blockHash)
-	if err != nil {
-		return "", err
-	}
-
-	if len(blockVb.Tx) <= 1 {
-		return "", fmt.Errorf("block %d has no transactions", blockVb.Height)
-	}
-
-	tss, err := ob.ZetacoreClient().GetBTCTSSAddress(ctx, ob.Chain().ChainId)
-	if err != nil {
-		return "", err
-	}
-
-	// check confirmation
-	// #nosec G115 block height always positive
-	if !ob.IsBlockConfirmed(uint64(blockVb.Height)) {
-		return "", fmt.Errorf("block %d is not confirmed yet", blockVb.Height)
-	}
-
-	// #nosec G115 always positive
-	event, err := GetBtcEvent(
-		ctx,
-		ob.rpc,
-		*tx,
-		tss,
-		uint64(blockVb.Height),
-		ob.logger.Inbound,
-		ob.netParams,
-		common.CalcDepositorFee,
-	)
-	if err != nil {
-		return "", err
-	}
-
-	if event == nil {
-		return "", errors.New("no btc deposit event found")
-	}
-
-	msg := ob.GetInboundVoteFromBtcEvent(event)
-	if msg == nil {
-		return "", errors.New("no message built for btc sent to TSS")
-	}
-
-	if !vote {
-		return msg.Digest(), nil
-	}
-
-	return ob.PostVoteInbound(ctx, msg, zetacore.PostVoteInboundExecutionGasLimit)
+	// successful processed all blocks in [startBlock, toBlock]
+	return toBlock, nil
 }
 
 // FilterAndParseIncomingTx given txs list returned by the "getblock 2" RPC command, return the txs that are relevant to us